refactor: rename `epoch_setting` record filename to `epoch_settings`
