+++ mithril-aggregator/src/database/record/mod.rs
@@ -2,7 +2,7 @@
 
 mod buffered_single_signature_record;
 mod certificate;
-mod epoch_setting;
+mod epoch_settings;
 mod open_message;
 mod open_message_with_single_signatures;
 mod signed_entity;
@@ -13,7 +13,7 @@ mod stake_pool;
 
 pub use buffered_single_signature_record::*;
 pub use certificate::*;
-pub use epoch_setting::*;
+pub use epoch_settings::*;
 pub use open_message::*;
 pub use open_message_with_single_signatures::*;
 pub use signed_entity::*;
